+++ option-parser.js
@@ -114,7 +114,7 @@ const config = {
 		enact = Object.assign(enact, meta);
 
 		// Parse the theme config tree for defaults
-		config.theme = themeConfig(pkg.path, process.env.ENACT_THEME || enact.theme || 'sandstone');
+		config.theme = themeConfig(pkg.path, process.env.ENACT_THEME || enact.theme || 'limestone');
 
 		// Optional alternate entrypoint for isomorphic builds.
 		config.isomorphic = computed('isomorphic', enact, config.theme);
@@ -145,7 +145,7 @@ const config = {
 		config.additionalModulePaths = computed('additionalModulePaths', enact, config.theme);
 		config.entry = computed('entry', enact, config.theme);
 
-		// Resolve array of screenType configurations. When not found, falls back to any theme preset or sandstone.
+		// Resolve array of screenType configurations. When not found, falls back to any theme preset or limestone.
 		const screens = computed('screenTypes', enact, config.theme);
 		config.screenTypes =
 			(Array.isArray(screens) && screens) ||
@@ -161,7 +161,7 @@ const config = {
 			? riConfig
 			: config.screenTypes.reduce((r, s) => (s.base && {baseSize: s.pxPerRem}) || r, undefined);
 
-		// Resolved filepath to fontGenerator. When not found, falls back to any theme preset or sandstone.
+		// Resolved filepath to fontGenerator. When not found, falls back to any theme preset or limestone.
 		const fontGenerator = computed('fontGenerator', enact, config.theme);
 		config.fontGenerator =
 			fontGenerator &&